now conditions should evaluate to 0/1

+++ TINY_Compiler_Scanner/JASON_Compiler/SyntaxAnalyser.cs
@@ -10,7 +10,7 @@ namespace JASON_Compiler
     public class Node
     {
         public List<Node> children = new List<Node>();
-        public int value = Int32.MinValue;
+        public float value = Int32.MinValue;
         public string datatype = "";
         public string Name;//to be deleted
         public Token token;
